notification sound, custom cell, and more
Added a notification sound to the UIAlertView when the goal is reached.
Called a UILocalNotification instead of a UIAlertView if the
application is in the background or the screen is locked. Added an
initial alert on first opening of the app to let the user know to
choose their units and weight in settings. Created a custom
UITableViewCell for workout info displaying on the main screen. Lots of
stuff accomplished!